+++ src/TestTrackingDiagrams.LightBDD.Core/StandardPipelineFormatter.cs
@@ -21,7 +21,7 @@ public class StandardPipelineFormatter : IReportFormatter
         if (ExpectedTestCount != null)
         {
             var scenarioCount = features.SelectMany(f => f.GetScenarios()).Count();
-            if (scenarioCount > ExpectedTestCount())
+            if (scenarioCount < ExpectedTestCount())
                 return;
         }